refactor: extract super ops sales actions
