+++ supabase/functions/super-ops/actions/index.ts
@@ -1,5 +1,6 @@
 import type { SuperOpsContext } from "../context.ts";
 import { handleControlCenterAction } from "./controlCenter.ts";
+import { handleSalesCustomersAction } from "./salesCustomers.ts";
 import { handleSecuritySessionsAction } from "./securitySessions.ts";
 import { handleSupportAction } from "./support.ts";
 
@@ -40,6 +41,7 @@ export const dispatchSuperOpsAction = async (
       handleSecuritySessionsAction,
       handleControlCenterAction,
       handleSupportAction,
+      handleSalesCustomersAction,
     ]
   ) {
     const response = await handler(context);
